Move tests from factories to their correct locations
